+++ Source/Private/common.ps1
@@ -37,7 +37,11 @@ function _callAPI {
       # that project name to be used in building the URI that would lead to
       # 404 because the URI would not be correct.
       [Alias('IgnoreDefaultProject')]
-      [switch]$NoProject
+      [switch]$NoProject,
+      # This flag makes sure that no specific account is used
+      # some APIs do not have an account in their API uri because
+      # they are not account specific in the url path itself. (e.g. user profile, pipeline billing)
+      [switch]$NoAccount
    )
 
    process {
@@ -95,8 +99,10 @@ function _callAPI {
       }
 
       # We have to remove any extra parameters not used by Invoke-RestMethod
-      $extra = 'NoProject', 'UseProjectId', 'Area', 'Resource', 'SubDomain', 'Id', 'Version', 'JSON', 'ProjectName', 'Team', 'Url', 'QueryString', 'AdditionalHeaders', 'CustomBearer'
-      foreach ($e in $extra) { $params.Remove($e) | Out-Null }
+
+      $extra = 'NoAccount', 'NoProject', 'UseProjectId', 'Area', 'Resource', 'SubDomain', 'Id', 'Version', 'JSON', 'ProjectName', 'Team', 'Url', 'QueryString', 'AdditionalHeaders', 'CustomBearer'
+
+   foreach ($e in $extra) { $params.Remove($e) | Out-Null }
 
       try {
          $resp = Invoke-RestMethod @params
@@ -132,12 +138,23 @@ function _testGraphSupport {
 function _supportsHierarchyQuery {
    _hasAccount
    if ($false -eq $(_testHierarchyQuerySupport)) {
-      throw 'This account does not support the graph API.'
+      throw 'This account does not support the hierarchy query API.'
    }
 }
 
 function _testHierarchyQuerySupport {
    (_getApiVersion HierarchyQuery) -as [boolean]
+}
+
+function _supportsBilling {
+   _hasAccount
+   if ($false -eq $(_testBillingSupport)) {
+      throw 'This account does not support the billing API.'
+   }
+}
+
+function _testBillingSupport {
+   (_getApiVersion Billing) -as [boolean]
 }
 
 function _supportVariableGroups {
@@ -184,7 +201,7 @@ function _getApiVersion {
          'DistributedTaskReleased', 'VariableGroups', 'Tfvc',
          'Packaging', 'MemberEntitlementManagement',
          'ExtensionsManagement', 'ServiceEndpoints', 'Graph',
-         'TaskGroups', 'Policy', 'Processes', 'HierarchyQuery', 'Pipelines')]
+         'TaskGroups', 'Policy', 'Processes', 'HierarchyQuery', 'Pipelines', 'Billing')]
       [string] $Service,
 
       [parameter(ParameterSetName = 'Target')]
@@ -238,7 +255,8 @@ function _buildRequestURI {
       $ProjectName,
 
       [switch]$UseProjectId,
-      [switch]$NoProject
+      [switch]$NoProject,
+      [switch]$NoAccount
    )
 
    process {
@@ -246,14 +264,19 @@ function _buildRequestURI {
 
       $sb = New-Object System.Text.StringBuilder
 
-      $sb.Append($(_addSubDomain -subDomain $subDomain -instance $(_getInstance))) | Out-Null
+      $instance = "https://dev.azure.com"
+      if ($NoAccount.IsPresent -eq $false) {
+         $instance = _getInstance
+      }
+
+      $sb.Append($(_addSubDomain -subDomain $subDomain -instance $instance)) | Out-Null
 
       # There are some APIs that must not have the project added to the URI.
       # However, if they caller set the default project it will be passed in
       # here and added to the URI by mistake. Functions that need the URI
       # created without the project even if the default project is set needs
       # to pass the -NoProject switch.
-      if ($ProjectName -and $NoProject.IsPresent -eq $false) {
+      if ($ProjectName -and $NoProject.IsPresent -eq $false -and $NoAccount.IsPresent -eq $false) {
          if ($UseProjectId.IsPresent) {
             $projectId = (Get-VSTeamProject -Name $ProjectName | Select-Object -ExpandProperty id)
             $sb.Append("/$projectId") | Out-Null
@@ -1017,12 +1040,21 @@ function _getBillingToken {
    # get a billing access token by using the given PAT.
    # this token can be used for buying pipelines or artifacts
    # or other things used for billing (except user access levels)
+   [CmdletBinding()]
+   param (
+      #billing token can have different scopes. They are defined by named token ids.
+      #They should be validated to be specific by it's name
+      [Parameter(Mandatory=$true)]
+      [string]
+      [ValidateSet('AzCommDeploymentProfile','CommerceDeploymentProfile')]
+      $NamedTokenId
+   )
 
    $sessionToken = @{
       appId        = 00000000 - 0000 - 0000 - 0000 - 000000000000
       force        = $false
       tokenType    = 0
-      namedTokenId = "CommerceDeploymentProfile"
+      namedTokenId = $NamedTokenId
    }
 
    $billingToken = _callAPI `